Slight Overhaul of UI
Made updates mainly to the resume and portfolio documents. Fixed incorrect icon images for linkedin and github on resume.

New streaming website wasn't displaying in the 2nd iframe, so it was replaced with a gallery of five images, including a link to the site.

An REU html doc was created to showcase the work done while during internship. It's been added to all nav menu's but is commented out until completed.