Update index.html and resume.css
The commit includes modifications to the structure and styles of the website. Changes in index.html mainly concern the reformatting of the carousel section. The major update in resume.css file focuses on improvements in website aesthetics such as font adjustments, color changes, and the addition of animations. Furthermore, it also optimizes the layout for different screen sizes.