Add collapse-arrows.js and update website layout
A new JavaScript file, 'collapse-arrows.js', has been added to control the collapse of elements in the resume section. This has led to updates in the layout in 'index.html' and 'resume.html' to include this new JavaScript file. Additionally, several changes have been made in the 'resume.css' file to improve the styling and layout.