Update CSS properties and hierarchy
This commit makes several updates to the resume.css file. It modifies the hierarchy for the 'section p' style, alters the display settings for '.up-arrow' and '.collapsed .down-arrow', and adds the '!important' rule to certain styles to override any conflicting declarations. Display property for '.truncate' was also changed to block.