Simplified Navbar, Updated SVG Icons, Updated resume content
Removed unnecessary SVG code in the navbar to improve page loading time. Additionally, eliminated the "Voted #1 Worldwide" text as it was found to be misleading to visitors. The modification enhances the website's usability and authenticity.

Removed the bs-icon-circle from the navbar-brand. Replaced the bell SVG icon in the project card with different SVG icons corresponding to the relevant subject of each card. The changes give better context and visual understanding of the content to the user.

Refactored the header to remove svg icon, included a new tab for certificates, and updated course titles. In the contacts section, extraneous spaces in names were removed and each contact given its own distinct comment. Lastly, significantly expanded the certificate section with additional course certification data.